added formatPhoneNumber utils
