+++ src/components/account/AccountSetting.jsx
@@ -1,17 +1,8 @@
 import { useState } from "react";
+import { formatPhoneNumber } from "../../utils/formatPhoneNumber.js";
 
 const AccountSetting = () => {
   const [phoneNumber, setPhoneNumber] = useState("");
-
-  const formatPhoneNumber = (input) => {
-    const numericInput = input.replace(/\D/g, "");
-    const limitedInput = numericInput.slice(0, 10);
-    const formattedInput = limitedInput.replace(
-      /^(\d{3})(\d{3})(\d{4})$/,
-      "($1) $2-$3"
-    );
-    return formattedInput;
-  };
 
   const handlePhoneNumberChange = (e) => {
     const input = e.target.value;
